requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.14.0
mysqlclient==2.2.4
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
